requirements.txt
python-dotenv~=1.1.0
discord.py~=2.5.2
aiohttp~=3.12.13